# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger",
        "description": "API CRUD Students",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get list of students",
                "description": "Return list of students",
                "operationId": "535d897b77e498b3fa28007dda371af6",
                "responses": {
                    "200": {
                        "description": "Succesful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StudentResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create new student",
                "description": "Create a new student record",
                "operationId": "cdcd906c4c8e328905218eec3971e8c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "age"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get student information",
                "description": "Get student details by ID",
                "operationId": "8f894dbfa733dcdb710bf599d3b95c36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update student information",
                "description": "Update student record by ID",
                "operationId": "d53e3ab7e094268789b668b20ee405a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "age"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Delete student record",
                "description": "Delete student by ID",
                "operationId": "9a3955a0d2e5ea12d45787f6ad608708",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StudentResource": {
                "properties": {
                    "id": {
                        "description": "Id of the student",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the student",
                        "type": "string"
                    },
                    "age": {
                        "description": "Age of the student",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Students",
            "description": "Students"
        }
    ]
}